pyproject.toml
[project]
name = "agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.2.0",
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.68",
    "langchain-groq>=0.3.5",
    "langchain-openai>=0.3.27",
    "langchain-tavily>=0.2.7",
    "langgraph>=0.5.1",
    "langgraph-cli[inmem]>=0.3.4",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.9",
    "wikipedia>=1.4.0",
]
